ud_project/saraswathi/requirements.txt
certifi==2024.2.2
click==8.1.7
cx_Oracle==8.3.0
databricks-sql-connector==3.1.0
diskcache==5.6.3
fastapi==0.110.0
httpx==0.27.2
matplotlib==3.8.3
mongomock==4.3.0
mysql-connector-python==8.3.0
numpy==1.26.4
openai==1.55.2
oracledb==2.2.1
package-obfuscator==1.0.0
pandas==2.1.4
pandasql==0.7.3
plotly==5.24.1
psycopg2-binary==2.9.9
pydantic==2.6.2
pydantic_core==2.16.3
Pygments==2.17.2
pymongo==4.10.1
PyMySQL==1.1.0
pyodbc==5.1.0
pysqlite3==0.5.2
pytest-cov==6.0.0
python-dotenv==1.0.1
python-Levenshtein==0.25.1
pyyaml==6.0.2
rapidfuzz==3.9.6
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
ruff==0.6.9
scipy==1.13.0
squarify==0.4.4
SQLAlchemy==2.0.29
starlette==0.36.3
tiktoken==0.6.0
trino==0.329.0
tqdm==4.66.2
uvicorn==0.27.1
qdrant-client==1.12.1